Slight restructuring in between the rap bits, also re-recorded bass in there, beefed up the drums in the build up from the sad bit.
Going from rap #2 into the chorus feels a bit disjointed atm, tried fixing with just bass but think might need vocal + guitar edits to get it feeling smooth.

Sad chordy bit sounds really nice, think it works well. At the moment though, it's not quite on beat if listening with the metronome on - having it a bit loose isn't necessarily a problem, but I could hear something was off with the metronome off. Maybe worth a re-record so it lines up better, but not super important.